+++ tests/TestCase.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Mcandylab\LaravelCuid2\Tests;
+
+use Mcandylab\LaravelCuid2\LaravelCuid2ServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app): array
+    {
+        return [
+            LaravelCuid2ServiceProvider::class,
+        ];
+    }
+
+    protected function defineEnvironment($app): void
+    {
+        $app['config']->set('database.default', 'testing');
+        $app['config']->set('database.connections.testing', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
+    }
+}